requirements.txt
python-barcode 
pillow 
pycups